+++ lib/utils/process_result_extension.dart
@@ -5,8 +5,8 @@ import 'package:process_run/src/lines_utils.dart';
 
 /// run response helper.
 extension ProcessRunProcessResultsExt on List<ProcessResult> {
-  Iterable<String> _outTextsToLines(Iterable<String> out) =>
-      out.expand((element) => LineSplitter.split(element));
+  Iterable<String> _outLinesToLines(Iterable<Iterable<String>> out) =>
+      out.expand((lines) => lines);
 
   /// Join the out lines for a quick string access.
   String get outText => outLines.join('\n');
@@ -16,11 +16,28 @@ extension ProcessRunProcessResultsExt on List<ProcessResult> {
 
   /// Out line lists
   Iterable<String> get outLines =>
-      _outTextsToLines(map((result) => result.stdout.toString()));
+      _outLinesToLines(map((result) => result.outLines));
 
   /// Line lists
   Iterable<String> get errLines =>
-      _outTextsToLines(map((result) => result.stderr.toString()));
+      _outLinesToLines(map((result) => result.errLines));
+}
+
+/// run response helper.
+extension ProcessRunProcessResultExt on ProcessResult {
+  Iterable<String> _outStringToLines(String out) => LineSplitter.split(out);
+
+  /// Join the out lines for a quick string access.
+  String get outText => outLines.join('\n');
+
+  /// Join the out lines for a quick string access.
+  String get errText => errLines.join('\n');
+
+  /// Out line lists
+  Iterable<String> get outLines => _outStringToLines(this.stdout.toString());
+
+  /// Line lists
+  Iterable<String> get errLines => _outStringToLines(this.stderr.toString());
 }
 
 /// Process helper.